+++ desktop/cmp/rest/RestGridModel.js
@@ -8,7 +8,7 @@ import {isFunction, isPlainObject} from 'lodash';
 import {GridModel} from '@xh/hoist/cmp/grid';
 import {HoistModel, managed, XH} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon/Icon';
-import {pluralize, throwIf, withDefault, apiDeprecated} from '@xh/hoist/utils/js';
+import {pluralize, throwIf, withDefault} from '@xh/hoist/utils/js';
 import {RestStore} from './data/RestStore';
 import {RestFormModel} from './impl/RestFormModel';
 
@@ -97,9 +97,9 @@ export class RestGridModel extends HoistModel {
     get store() {return this.gridModel.store}
     /** @return {StoreSelectionModel} */
     get selModel() {return this.gridModel.selModel}
-    /** @return {Record[]} */
+    /** @return {StoreRecord[]} */
     get selectedRecords() {return this.gridModel.selectedRecords}
-    /** @return {Record} */
+    /** @return {StoreRecord} */
     get selectedRecord() {return this.gridModel.selectedRecord}
 
     /**
@@ -176,12 +176,6 @@ export class RestGridModel extends HoistModel {
     /** Load the underlying store. */
     loadData(...args) {
         return this.store.loadData(...args);
-    }
-
-    /** @deprecated */
-    get selection() {
-        apiDeprecated('RestGridModel.selection', {msg: 'Use selectedRecords instead', v: 'v44'});
-        return this.selectedRecords;
     }
 
     //-----------------