+++ CHANGELOG.md
@@ -9,8 +9,23 @@
 ### 💥 Breaking Changes
 
 * Mobile `Button` no longer supports `modifier` prop. Use `minimal` and `outlined` instead.
+* `Record` class renamed to `StoreRecord` in anticipation of upcoming changes to JavaScript
+  standard and to improve compatibility with TypeScript.
+* The following deprecated APIs were removed:
+  * GridModel.selection
+  * GridModel.selectedRecordId
+  * StoreSelectionModel.records
+  * StoreSelectionModel.ids
+  * StoreSelectionModel.singleRecord
+  * StoreSelectionModel.selectedRecordId
+  * DataViewModel.selection
+  * DataViewModel.selectedRecordId
+  * RestGridModel.selection
+  * LogUtils.withShortDebug
+  * Promise.start
 
 * [Commit Log](https://github.com/xh/hoist-react/compare/v44.3.0...develop)
+
 
 ## v44.3.0 - 2021-12-15
 
@@ -66,6 +81,7 @@
 * Desktop `Select.hideDropdownIndicator` now defaults to `true` on tablet devices due to UX bugs
   with the select library component and touch devices.
 * Ensure `Column.autosizeBufferPx` is respected if provided.
+
 
 ### ✨ Style